// src/packages/bui-react/src/components/_base/Flyout/Flyout.types.ts
type WithControlled = { active: boolean; defaultActive?: never };

type BaseProps = {
  id?: string;
  triggerType?: "hover" | "click" | "focus";
  position?: FlyoutPosition;
  forcePosition?: boolean;
  trapFocusMode?: TrapMode;
  navigationMode?: NavigationMode;
  children?: React.ReactNode;
  onOpen?: () => void;
  onClose?: () => void;
  onAfterOpen?: (e: React.TransitionEvent | TransitionEvent) => void;
  onAfterClose?: (e: React.TransitionEvent | TransitionEvent) => void;
  zIndex?: number | string;
  contentClassName?: string;
  contentAttributes?: G.Attributes<"div">;
  follow?: boolean;
  disableAnimation?: boolean | "hide";
  keepMounted?: boolean;
  timeout?: "short" | "long";
  triggerRef?: React.RefObject<HTMLElement>;
  positionRef?: React.RefObject<HTMLElement>;
  containerRef?: React.RefObject<HTMLElement>;
};

export type DefaultProps = Required<{
  position: BaseProps["position"];
  trigger: BaseProps["triggerType"];
}>;

export type UncontrolledProps = BaseProps & WithUncontrolled;
export type ControlledProps = BaseProps & WithControlled;
export type Props = ControlledProps | UncontrolledProps;
// ...
export type ContextProps = {
  id: string;
  flyout: ReturnType<UseFlyout>;
  flyoutTriggerRef: React.RefObject<HTMLElement>;
  flyoutContainerRef?: React.RefObject<HTMLElement>;
  flyoutElRef: React.RefObject<HTMLDivElement>;
  flyoutArrowElRef?: React.RefObject<HTMLElement>;
  handleOpen: (e: React.FocusEvent | FocusEvent) => void;
  handleClose: (e: React.FocusEvent | FocusEvent) => void;
  handleFocus: (e: React.FocusEvent | FocusEvent) => void;
  handleBlur: (e: React.FocusEvent | FocusEvent) => void;
  handleMouseEnter: (e: React.MouseEvent | MouseEvent) => void;
  handleMouseLeave: (e: React.MouseEvent | MouseEvent) => void;
  handleMouseMove: (e: React.MouseEvent | MouseEvent) => void;
  handleClick: (e: React.MouseEvent | MouseEvent) => void;
  handleTransitionEnd: (e: React.TransitionEvent | TransitionEvent) => void;
  follow?: boolean;
  focusTrapped: boolean;
} & Pick<
  Props,
  | "triggerType"
  | "contentClassName"
  | "contentAttributes"
  | "disableAnimation"
  | "trapFocusMode"
  | "navigationMode"
>;

export type RefProps = { open: () => void; close: () => void };
export type Ref = React.Ref<RefProps>;

export type Compound = React.ForwardRefExoticComponent<
  Props & { ref?: Ref }
> & {
  Trigger: React.ComponentType<TriggerProps>;
  Content: React.ComponentType<ContentProps>;
  Arrow: React.ComponentType<ArrowProps>;
};

/* Utilities */
type ElementRef = React.RefObject<HTMLElement | null>;

type PassedFlyoutOptions = {
  zIndex?: number | string;
  position?: FlyoutPosition;
  defaultActive?: boolean;
  forcePosition?: boolean;
  keepMounted?: boolean;
  relative?: boolean;
  flyoutArrowElRef: ElementRef;
  flyoutContainerRef?: ElementRef;
  getPositionCompensation?: (
    position: FlyoutPosition,
    trigger: HTMLElement | null,
    arrow: HTMLElement | null
  ) => FlyoutPositionCompensation | undefined;
};

export type FlyoutOptions = Pick<
  PassedFlyoutOptions,
  "forcePosition" | "zIndex"
> & {
  position: FlyoutPosition;
  rtl: boolean;
  mobileFallback?: boolean;
};

export type FlyoutStyles = React.CSSProperties;
export type FlyoutState = {
  styles: FlyoutStyles;
  position: FlyoutPosition;
  active: boolean;
  visible: boolean;
};
// ...
export type UseFlyout = (
  originRef: ElementRef,
  targetRef: ElementRef,
  options: PassedFlyoutOptions
) => Pick<FlyoutState, "styles" | "position" | "active" | "visible"> & {
  render: () => void;
  show: () => void;
  hide: () => void;
  remove: () => void;
  update: () => void;
  move: (targetX: number, targetY: number) => void;
};